Fix all Kotlin compilation warnings in libapp module
Fixed type safety issues, null checks, and experimental API usage:
- GuideInterface: Fixed String? to String conversion with null assertion
- RingBuffer: Added null checks for ByteArray? parameters in System.arraycopy
- UsbBuffer: Removed redundant null checks on non-null properties
- ByteUtils: Added @OptIn annotation for ExperimentalUnsignedTypes
- FileUtils: Added null safety for File.listFiles() return value
- Applied same fixes to duplicate files in mpdc4gsr package

+++ libapp/src/main/java/com/matrix/utils/FileUtils.kt
@@ -41,16 +41,18 @@ class FileUtils {
             }
             flag = true
             val files = dirFile.listFiles()
+            
+            if (files != null) {
+                for (i in files.indices) {
+                    if (files[i].isFile) {
 
-            for (i in files.indices) {
-                if (files[i].isFile) {
+                        flag = deleteFile(files[i].absolutePath)
+                        if (!flag) break
+                    } else {
 
-                    flag = deleteFile(files[i].absolutePath)
-                    if (!flag) break
-                } else {
-
-                    flag = deleteDirectory(files[i].absolutePath)
-                    if (!flag) break
+                        flag = deleteDirectory(files[i].absolutePath)
+                        if (!flag) break
+                    }
                 }
             }
             return if (!flag) false else dirFile.delete()
